Update login flow
