+++ backend/src/routes/auth.ts
@@ -63,7 +63,7 @@ router.get('/google/callback', async (req, res) => {
     });
 
     const token = authService.generateToken(user);
-    const frontendUrl = new URL('http://localhost:3000/Home');
+    const frontendUrl = new URL('http://localhost:3000/Login');
     frontendUrl.searchParams.set('token', token);
     frontendUrl.searchParams.set('user', JSON.stringify({
         id: user._id?.toString(),
